asm-verify #194 Wave B: refactor 8 free fns to the binary signature (pair)
8 ported free functions paired under a different signature than the binary, so
asm-verify still couldn't diff them. Refactored to the binary's exact symbol
(all binary-faithful, behavior-preserving):
- StringHash: two-overload -- out-of-line StringHash(const char*, int len) (the
  int is the byte length, per binary @0x00252a10) + inline 1-arg forwarder.
  Emits both _Z10StringHashPKci and _Z10StringHashPKc; 229 call sites unchanged.
- GameInitialise: restore (void* window, const char* config) @0x0011d22c (args
  unused -- SDL owns its window; // DIFFERS); callers pass nullptr,nullptr.
- CreateEntity: int -> long @0x001d8fec (+ TODO for the unported case-6 FruitRay).
- GetBonusTotal: drop static, uint64_t -> unsigned long @0x0012e4c4.
- SetCurrentModeHighscore: free bool(int) wrapper @0x00119f24 over the save
  singleton (member kept).
- MoveFruitZPositionToBack: de-method, Fruit* -> float& @0x001ca674.
- GetComboStarText/Texture: de-method + uint8_t -> COMBO_TYPE enum (25 values)
  @0x001325f8 / @0x00132a94.
- IsFastHardware: drop static @0x0011f394.
- CriticalFlash: already free + const-ref -> pairs as-is (verify only).
AddSlice deferred to the faithful SliceEffect subsystem port (#196), not a rename.
Build green; 20/20 tests pass.

+++ src/entities/ActorManager.h
@@ -44,7 +44,7 @@ public:
     static const int FREE_POOL_CAP = 512;
 
     // Factory delegate signature — matches Mortar::Delegate1<Entity*, long>.
-    typedef Entity* (*FactoryFn)(int entityType);
+    typedef Entity* (*FactoryFn)(long entityType);
 
     // --- Fields mirrored from binary layout (sizes/offsets in comments) -
